debug component wip
