+++ src/component/debug/BuildInfo/index.js
@@ -1,2 +1,1 @@
-export {default} from "./info";
-export {COMPONENT_ID} from "./info.relocation";
+export {default} from './BuildInfo';
